+++ lib/chartmogul/subscription_event.rb
@@ -6,6 +6,8 @@ module ChartMogul
     set_resource_path '/v1/subscription_events'
 
     readonly_attr :id
+    readonly_attr :disabled
+    readonly_attr :disabled_at
     writeable_attr :data_source_uuid
     writeable_attr :customer_external_id
     writeable_attr :subscription_set_external_id
@@ -49,10 +51,70 @@ module ChartMogul
                                 :subscription_event)
     end
 
+    # Instance method: destroys this subscription event
     def destroy!
       handling_errors do
         connection.delete(resource_path.path, subscription_event: { id: instance_attributes[:id] })
       end
+    end
+
+    # Class method: accepts both flat and envelope-wrapped params for backwards compatibility
+    #   flat params.           : SubscriptionEvent.destroy!(id: 123)
+    #   envelope-wrapped params: SubscriptionEvent.destroy!(subscription_event: { id: 123 })
+    def self.destroy!(params = {})
+      body = params.key?(:subscription_event) ? params : { subscription_event: params }
+      handling_errors do
+        connection.delete(resource_path.path, body)
+      end
+      true
+    end
+
+    # Toggle the disabled state of a subscription event.
+    # Note: SubscriptionEvent uses envelope-wrapped requests/responses ({ subscription_event: ... })
+    # unlike other resources, because the API requires this format for this endpoint.
+    def toggle_disabled!(disabled:, handle_as_user_edit: nil)
+      path = self.class.build_query_path(
+        "#{resource_path.path}/#{instance_attributes[:id]}/disabled_state",
+        handle_as_user_edit: handle_as_user_edit
+      )
+      resp = handling_errors { self.class.json_patch(path, { disabled: disabled }) }
+      json = ChartMogul::Utils::JSONParser.parse(resp.body, immutable_keys: self.class.immutable_keys)
+      assign_all_attributes(json[:subscription_event] || json)
+      self
+    end
+
+    # Update a subscription event by data_source_uuid and external_id.
+    # Uses envelope-wrapped body ({ subscription_event: ... }) as required by this API endpoint.
+    def self.update_by_external_id!(data_source_uuid:, external_id:, handle_as_user_edit: nil, **attributes)
+      path = build_query_path(resource_path.path, handle_as_user_edit: handle_as_user_edit)
+      resp = handling_errors do
+        json_patch(path, { subscription_event: attributes.merge(data_source_uuid: data_source_uuid, external_id: external_id) })
+      end
+      json = ChartMogul::Utils::JSONParser.parse(resp.body, immutable_keys:)
+      new_from_json(json[:subscription_event] || json)
+    end
+
+    # Delete a subscription event by data_source_uuid and external_id
+    def self.destroy_by_external_id!(data_source_uuid:, external_id:, handle_as_user_edit: nil)
+      path = build_query_path(resource_path.path, handle_as_user_edit: handle_as_user_edit)
+      handling_errors do
+        connection.delete(path, subscription_event: { data_source_uuid: data_source_uuid, external_id: external_id })
+      end
+      true
+    end
+
+    # Toggle disabled state of a subscription event by data_source_uuid and external_id.
+    # Uses envelope-wrapped body ({ subscription_event: ... }) as required by this API endpoint.
+    def self.toggle_disabled_by_external_id!(data_source_uuid:, external_id:, disabled:, handle_as_user_edit: nil)
+      path = build_query_path("#{resource_path.path}/disabled_state", handle_as_user_edit: handle_as_user_edit)
+      resp = handling_errors do
+        json_patch(path, {
+                     subscription_event: { data_source_uuid: data_source_uuid, external_id: external_id },
+                     disabled: disabled
+                   })
+      end
+      json = ChartMogul::Utils::JSONParser.parse(resp.body, immutable_keys:)
+      new_from_json(json[:subscription_event] || json)
     end
   end